\name{internal-functions}
\alias{startZ}
\alias{startLogit}
\alias{startSend}
\alias{startRec}
\alias{startLambda}
\alias{showTrace}
\alias{list2array}
\alias{sigmaFC}
\alias{muFc}
\alias{alphaBetaProp_k}
\alias{zProp_i}
\alias{lambdaProp_l}
\alias{gammaThetaProp_i}
\alias{logPosterior}
\alias{loglik}


\title{
Internal \code{spaceNet} functions
}
\description{
Internal functions not to be called by the user.
}

\keyword{internal}
